requirements-dev.txt
-r requirements.txt
pytest>=8.0.0
pytest-homeassistant-custom-component>=0.13.109
pytest-asyncio
pytest-cov
sqlalchemy
black
isort
flake8
pylint
mypy
